Moved ArtifactLocation to a separate file
